requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
docker==7.1.0
docutils==0.21.2
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown2==2.5.0
MarkupSafe==2.1.5
multidict==6.1.0
nicegui==2.1.0
orjson==3.10.7
pscript==0.7.7
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.10
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vbuild==0.8.2
watchfiles==0.24.0
websockets==13.1
wsproto==1.2.0
yarl==1.12.1
mntfinder==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-ldap==3.4.4
mount==1.0.0